% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export_donnees_sd_atlas.R
\name{export_donnees_sd_atlas}
\alias{export_donnees_sd_atlas}
\title{Exporter au format csv le tableau de données atlas mis à jour}
\usage{
export_donnees_sd_atlas(df, fichier_csv_sortie = NA)
}
\arguments{
\item{df}{Le dataframe contenant les données mises à jour.}

\item{fichier_csv_sortie}{Caractère. Chemin vers le fichier csv de sortie. S'il n'est pas précisé,
la fonction va nommer le fichier CSV d'après les date et heure du système et l'écrire
dans un sous-répertoire "processed_data" du répertoire du travail.}
}
\value{
Un fichier csv au format souhaité.
}
\description{
Exporter au format csv le tableau de données atlas mis à jour
}
\examples{
\dontrun{
export_donnees_sd_atlas(
  df = data,
  fichier_csv_sortie = "donnees_traitees/2020_08_16_peches_sd_ofb_atlas.csv"
)
}
}
