% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maj_donnees_sd_atlas.R
\name{maj_donnees_sd_atlas}
\alias{maj_donnees_sd_atlas}
\title{Mettre à jour les données de pêches SD OFB pour la plateforme régionale des données naturalistes}
\usage{
maj_donnees_sd_atlas(df_a_jour, df_precedent)
}
\arguments{
\item{df_a_jour}{Nom du dataframe contenant les données à jour.}

\item{df_precedent}{Nom du dataframe contenant les données de la précédente livraison
de données à la plateforme.}
}
\value{
Le dataframe fusionné avec les données au format souhaité et les bons UUID.
}
\description{
La fonction compare une version à jour du tableau de données mis en forme par la fonction mef_donnees_sd_atlas()
    avec une version antérieure préparée avec la même fonction.
}
\examples{
\dontrun{
data <- maj_donnees_sd_atlas(
  fichier_shp_a_jour = "donnees_brutes/atlas.shp",
  fichier_csv_precedent = "donnees_traitees/2020_08_16_peches_sd_ofb_atlas.csv"
)
}
}
