% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mef_donnees_sd_atlas.R
\name{mef_donnees_sd_atlas}
\alias{mef_donnees_sd_atlas}
\title{Mettre en former les données de pêches SD OFB pour la plateforme régionale des données naturalistes}
\usage{
mef_donnees_sd_atlas(fichier_shp_a_jour, crs_init = 2154, crs_fin = 4326)
}
\arguments{
\item{fichier_shp_a_jour}{Caractère. Chemin vers le fichier shapefile contenant les données à jour.}

\item{crs_init}{Numérique. Code EPSG du CRS initial. Par défaut c'est 2154 (Lambert 93).}

\item{crs_fin}{Numérique. Code EPSG du CRS de sortie. Par défaut c'est 4326 (WGS84).}
}
\value{
Le dataframe avec les données au format souhaité (long, nommage et typage des variables).
}
\description{
La fonction compare une version à jour du ficj=hier de données QG
}
\examples{
\dontrun{
data <- mef_donnees_sd_atlas(
  fichier_shp_a_jour = "donnees_brutes/atlas.shp")
}
}
