% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maj_donnees_sd_atlas.R
\name{maj_donnees_sd_atlas}
\alias{maj_donnees_sd_atlas}
\title{Mettre à jour les données de pêches SD OFB pour la plateforme régionale des données naturalistes}
\usage{
maj_donnees_sd_atlas(
  fichier_shp_a_jour,
  fichier_csv_precedent,
  fichier_csv_a_jour
)
}
\arguments{
\item{fichier_shp_a_jour}{Caractère. Chemin vers le fichier shapefile contenant les données à jour.}

\item{fichier_csv_precedent}{Caractère. Chemin vers le fichier csv de la précédente livraison
de données à la plateforme.}

\item{fichier_csv_a_jour}{Caractère. Chemin vers le fichier csv à jour.}
}
\value{
Le dataframe avec les données au format souhaité (long, nommage et typage des variables).
}
\description{
La fonction compare une version à jour du ficj=hier de données QG
}
\examples{
\dontrun{
maj_donnees_sd_atlas(
  fichier_shp_a_jour = "donnees_brutes/atlas.shp",
  fichier_csv_precedent = "donnees_traitees/2020_08_16_peches_sd_ofb_atlas.csv",
  fichier_csv_a_jour = "donnees_traitees/2021_08_16_peches_sd_ofb_atlas.csv"
)
}
}
